Exercício 07 - maior média entre as matérias
